pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "day_ahead_v2"
version = "2.0.0"
description = ""
authors = [
  { name = "yahei-DTU", email = "yahei@dtu.dk" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openmeteo-requests>=1.7.5",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "tabulate>=0.9.0",
    "torch==2.5.1+cpu",
    "torchvision==0.20.1+cpu",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.14.11",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
